pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "langfuse>=3.12.0",
    "pydantic-ai>=1.44.0",
    "python-dotenv>=1.2.1",
    "temporalio>=1.20.0",
    "uvicorn>=0.40.0",
]
